Demonstrate notebook viewing page w query params
This demonstrates a dynamically router page for individual notebooks.
The notebook name is a variable that is part of the URL. The page also
takes the query parameters as initialStaticProps.

To read the query parameters, we use a custom version of the useRouter
hook; see https://github.com/vercel/next.js/issues/4804#issuecomment-606624325