Add mock API for times-square rendered HTML
This demonstrates adding an iframe with a src attribute that points to
the times-square v1 API endpoint for the rendered HTML of a page.

In the next config, this URL is re-written to a path serviced by the
next.js node server to mock out the response; in real-life the
Kubernetes ingress would route this to the times-square API service.