Implement auto-refreshing notebook
Times Square API provides a new /htmlstatus endpoint that provides
metadata about the HTML render needed to support auto-refreshing the
content: a flag of whether the HTML is available or not, and a digest of
the HTML to indicate whether the client has the same version of the HTML
as is on the server (i.e. to support notebooks with TTL settings). If
the notebook is not available yet, this is set to a default value. Since
swr revalidates data (every 1 second right now) the default key will be
replaced with the digest. Changing this key on the iframe causes the
iframe to reload.

Note that useHtmlStatus is refactored into a reusable hook; other
components can also use html status and share the same connection pool
(so no additional requests are being made).